requirements.txt
cachetools==5.3.1
django==4.2.6
django-cors-headers==4.3.0
django-db-geventpool==4.0.1
django-debug-toolbar
django-debug-toolbar-force
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
docutils==0.20.1
drf-yasg[validation]==1.21.7
gunicorn[gevent]==21.2.0
hexbytes==0.3.1
packaging>=21.0
psycogreen==1.0.2
psycopg2==2.9.9
requests==2.31.0
safe-eth-py[django]==6.0.0b7
web3==6.11.1
